f/mmcdaniel-RASS: Add optional AdminTime parameter to BolusAction
Give the option to specify the length of time over which a bolus will be
administered.  Default time of 2 s was unrealistically fast and was
causing peak plasma concentrations for most bolus drugs to be too high.

Increase default admin time to 10 s and overwrite with user provided
admin time if provided.